+++ api/src/test/java/org/grnet/cat/api/registry/TestEndpointTest.java
@@ -8,12 +8,14 @@ import org.grnet.cat.api.KeycloakTest;
 import org.grnet.cat.api.endpoints.registry.TestEndpoint;
 import org.grnet.cat.dtos.AutomatedTestDto;
 import org.grnet.cat.dtos.InformativeResponse;
+import org.grnet.cat.dtos.registry.test.CheckUrlRequestDto;
 import org.grnet.cat.dtos.registry.test.TestRequestDto;
 import org.grnet.cat.dtos.registry.test.TestResponseDto;
 import org.grnet.cat.dtos.registry.test.TestUpdateDto;
 import org.junit.jupiter.api.Test;
 
 import static io.restassured.RestAssured.given;
+import static io.restassured.RestAssured.urlEncodingEnabled;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @QuarkusTest
@@ -209,12 +211,14 @@ public class TestEndpointTest extends KeycloakTest {
     public void testValidHttpsUrl() {
         register("admin");
 
+        var request=new CheckUrlRequestDto();
+        request.url="https://google.com";
         var response = given()
                 .auth()
                 .oauth2(getAccessToken("admin"))
+                .body(request)
                 .contentType(ContentType.JSON)
-                .queryParam("url", "https://google.com")
-                .get("/check-automated-test/https-url")
+                .post("/check-automated-test/https-url")
                 .then()
                 .assertThat()
                 .statusCode(200)
@@ -226,33 +230,77 @@ public class TestEndpointTest extends KeycloakTest {
     }
 
     @Test
-    public void testInvalidHttpUrl() {
+    public void testInvalidHttpsUrl() {
         // Mock the response for invalid URL
         register("admin");
+        var request=new CheckUrlRequestDto();
+        request.url="https://google1.com";
 
         var error = given()
                 .auth()
                 .oauth2(getAccessToken("admin"))
+                .body(request)
                 .contentType(ContentType.JSON)
-                .queryParam("url", "http://example.com")
-                .get("/check-automated-test/https-url")
+                .post("/check-automated-test/https-url")
                 .then()
                 .assertThat()
                 .statusCode(400)
                 .extract()
                 .as(InformativeResponse.class);
-        assertEquals("Failed to connect to the URL: http://example.com. The URL is not a secure HTTPS connection.", error.message);
+        assertEquals("Failed to connect to the URL: "+request.url+". IOException: "+request.url.replace("https://",""), error.message);
     }
     @Test
-    public void testUnauthenticatedUser() {
-        // register("alice");
+    public void testEmptyUrl() {
+        // Mock the response for invalid URL
+        register("admin");
+        var request=new CheckUrlRequestDto();
 
         var error = given()
                 .auth()
                 .oauth2(getAccessToken("admin"))
+                .body(request)
                 .contentType(ContentType.JSON)
-                .queryParam("url", "https://google.com")
-                .get("/check-automated-test/https-url")
+                .post("/check-automated-test/https-url")
+                .then()
+                .assertThat()
+                .statusCode(400)
+                .extract()
+                .as(InformativeResponse.class);
+        assertEquals("url may not be empty.", error.message);
+    }
+
+    @Test
+    public void testNoHttpsUrl() {
+        // Mock the response for invalid URL
+        register("admin");
+        var request=new CheckUrlRequestDto();
+        request.url="http://google.com";
+
+        var error = given()
+                .auth()
+                .oauth2(getAccessToken("admin"))
+                .body(request)
+                .contentType(ContentType.JSON)
+                .post("/check-automated-test/https-url")
+                .then()
+                .assertThat()
+                .statusCode(400)
+                .extract()
+                .as(InformativeResponse.class);
+        assertEquals("Failed to connect to the URL: "+request.url+". The URL is not a secure HTTPS connection.", error.message);
+    }
+    @Test
+    public void testUnauthenticatedUser() {
+        // register("alice");
+        var request=new CheckUrlRequestDto();
+        request.url="https://google.com";
+
+        var error = given()
+                .auth()
+                .oauth2(getAccessToken("admin"))
+                .body(request)
+                .contentType(ContentType.JSON)
+                .post("/check-automated-test/https-url")
                 .then()
                 .assertThat()
                 .statusCode(403)
@@ -264,13 +312,15 @@ public class TestEndpointTest extends KeycloakTest {
     @Test
     public void testForbiddenUser() {
         register("alice");
+        var request=new CheckUrlRequestDto();
+        request.url="https://google.com";
 
         var error = given()
                 .auth()
                 .oauth2(getAccessToken("admin"))
                 .contentType(ContentType.JSON)
-                .queryParam("url", "https://google.com")
-                .get("/check-automated-test/https-url")
+                .body(request)
+                .post("/check-automated-test/https-url")
                 .then()
                 .assertThat()
                 .statusCode(403)